api/requirements.txt
fastapi
uvicorn[standard]
pymongo
mysql-connector-python
requests